BDA,BDD,CSV,CL,DCL,

+++ Entities/CSV.cs
@@ -12,25 +12,25 @@ namespace CBM_API.Entities
         public int DeviceId { get; set; }
         
         public DateTime DateTest { get; set; }
-        public float? Outside { get; set; }
-        public float? Temperature { get; set; }
-        public float? Pd { get; set; }
-        public float? PdOnline { get; set; }
-        public float? HistoryMain { get; set; }
-        public float? NumberYearOper { get; set; }
-        public float? RIsolate { get; set; }
-        public float? PdByIndeSource { get; set; }
-        public float? PowerK { get; set; }
-        public float? PdAnalysis { get; set; }
-        public float? ScoreLevel1 { get; set; }
-        public float? ScoreLevel23 { get; set; }
-        public float? TotalScore { get; set; }
+        public double? Outside { get; set; }
+        public double? Temperature { get; set; }
+        public double? Pd { get; set; }
+        public double? PdOnline { get; set; }
+        public double? HistoryMain { get; set; }
+        public double? NumberYearOper { get; set; }
+        public double? RIsolate { get; set; }
+        public double? PdByIndeSource { get; set; }
+        public double? PowerK { get; set; }
+        public double? PdAnalysis { get; set; }
+        public double? ScoreLevel1 { get; set; }
+        public double? ScoreLevel23 { get; set; }
+        public double? TotalScore { get; set; }
         public string? Note { get; set; }
         public string? ReviewETC { get; set; }
         public string? Img { get; set; }
         public Device? Device { get; set; }
 
-        public CSV(int id, int deviceId, DateTime dateTest, float? outside, float? temperature, float? pd, float? pdOnline, float? historyMain, float? numberYearOper, float? rIsolate, float? pdByIndeSource, float? powerK, float? pdAnalysis, float? scoreLevel1, float? scoreLevel23, float? totalScore, string? note, string? reviewETC, string? img, Device? device)
+        public CSV(int id, int deviceId, DateTime dateTest, double? outside, double? temperature, double? pd, double? pdOnline, double? historyMain, double? numberYearOper, double? rIsolate, double? pdByIndeSource, double? powerK, double? pdAnalysis, double? scoreLevel1, double? scoreLevel23, double? totalScore, string? note, string? reviewETC, string? img, Device? device)
         {
             Id = id;
             DeviceId = deviceId;
